requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
openai==1.3.0
python-dotenv==1.0.0
pydantic[email]
reportlab==4.0.7
lxml==4.9.3
sqlalchemy==2.0.23
requests==2.31.0
cryptography==41.0.7
pydantic-settings==2.13.1
pytest
